+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Current
 -------
 
 - Fix missing changelog inprevious release
+- Ensure definitions with both `$ref` and description (or other property) output is valid (using `allOf`)
 
 0.12.0 (2018-09-27)
 -------------------